backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
python-multipart==0.0.20
aiofiles==24.1.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4

